- fix build errors under windows/msvc2010 - add platform check for intrin.h and stdint.h

+++ snd_blargg/boost/cstdint.hpp
@@ -27,8 +27,8 @@ typedef short           int16_t;
 typedef unsigned short  uint16_t;
 
 #if ULONG_MAX == 0xFFFFFFFF
-	typedef long            int32_t;
-	typedef unsigned long   uint32_t;
+	typedef int             int32_t;
+	typedef unsigned int    uint32_t;
 #elif UINT_MAX == 0xFFFFFFFF
 	typedef int             int32_t;
 	typedef unsigned int    uint32_t;
